MAKE IT LOOK GOOD?

+++ export/index.php
@@ -1,7 +1,23 @@
+<!DOCTYPE html>
+<html>
+	<head>
+		<meta charset="UTF-8">
+		<meta name="author" content="Caleb Milligan">
+		<title>Export data</title>
+		<link rel="stylesheet" type="text/css" href="../css/scouting.css">
+		<link rel="stylesheet" type="text/css" href="../css/errpage.css">
+	</head>
+	<body>
+		<p>Please wait while your download is prepared</p>
+		<div class="image-center">
+			<img width="64" height="64" src="../images/loading_icon.gif">
+		</div>
+	</body>
+</html>
 <?php
 $file = uniqid("", true) . ".sql";
 $output = `mysqldump -u username -ppassword --no-create-info --skip-triggers database pit_scouting stand_scouting total_points > $file`;
-if (file_exists($file)) {
+if (file_exists($file) && filesize($file) > 0) {
 	header('Content-Description: File Transfer');
 	header('Content-Type: application/octet-stream');
 	header('Content-Disposition: attachment; filename=scoutdata.sql');
@@ -13,13 +29,14 @@ if (file_exists($file)) {
 	ob_clean();
 	flush();
 	readfile($file);
-	unlink($file);
-	exit;
+	if (file_exists($file)) {
+		unlink($file);
+	}
+}
+else {
+	if (file_exists($file)) {
+		unlink($file);
+	}
+	header("Location: ../errpage.php");
 }
 ?>
-<!DOCTYPE html>
-<html>
-	<head>
-		<script type="text/javascript">window.close()</script>
-	</head>
-</html>
